requirements.txt
# Các thư viện chính
fastapi>=0.115.12
uvicorn>=0.34.0
pydantic>=2.11.0
starlette>=0.46.1
pydantic-settings>=2.8.1
pydantic-extra-types>=2.10.3
python-multipart>=0.0.20
email-validator>=2.2.0
ujson>=5.10.0
orjson>=3.10.16
itsdangerous>=2.2.0
python-dotenv>=1.0.0

# Các thư viện xử lý dữ liệu và NLP
transformers>=4.20.0
torch>=1.10.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0

# Các thư viện phục vụ tải file và xử lý dữ liệu
requests>=2.25.0
aiofiles>=0.8.0
openpyxl>=3.0.0
xlrd>=2.0.0

# Các thư viện AI
openai>=1.10.0
google-generativeai>=0.4.0

#Python 3.10
#Cuda -v 11.8
#Cudnn -v 8.5
